Add complex flux to demonstrate the ETL capacity
The flux declares to read an elasticsearch index, taking the AlephMabXml
stored in the elasticsearch, transform that into ntriples and load that to
into a mysql db. In parallel, a file is streamed to the harddisk and a
statistic of transformed fields is generated.
The example is taken from the lobid/lodmill repo and a Readme.textile
comes with it.